@@ -7,12 +7,12 @@
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.1.3",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-bootstrap": "^2.2.3",
     "react-burger-menu": "^2.7.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
+    "react-scripts": "^2.1.3",
     "react-slick": "^0.28.1",
     "slick-carousel": "^1.8.1",
     "web-vitals": "^2.1.4"
